requirements.txt
# Core ML and Data Processing
torch>=2.0.0
transformers>=4.30.0
datasets>=2.14.0
accelerate>=0.20.0
peft>=0.4.0

# Data Processing and Storage
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=12.0.0
fastparquet>=2023.4.0

# Deduplication (SlimPajama)
datasketch>=1.6.0
scikit-learn>=1.3.0

# Cloud Storage
boto3>=1.28.0
s3fs>=2023.6.0
gcsfs>=2023.6.0
google-cloud-storage>=2.10.0
azure-storage-blob>=12.18.0

# Evaluation Metrics
sacrebleu>=2.3.0
rouge-score>=0.1.2
bert-score>=0.3.13

# Korean NLP
mecab-python3>=1.0.6
sentencepiece>=0.1.99

# Utilities and Monitoring
tqdm>=4.65.0
psutil>=5.9.0
pyyaml>=6.0
click>=8.1.0
python-dotenv>=1.0.0
ray==2.31.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Jupyter for analysis notebooks
jupyter>=1.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Tokenization
tokenizers>=0.13.0
smart_open>=6.4.0
redis>=5.0.0
fakeredis>=2.23.0
mlflow>=2.12.0
